Build deterministic assembly foundation
Model settings: GPT-5 (default)

Thread: 019fc402-179a-7f41-ba4d-e89ea2646158

+++ src/legolization/stability/model.py
@@ -267,7 +267,7 @@ def footprint_columns(
 def knob_pattern(
     columns: frozenset[tuple[int, int]],
     min_dim: int,
-    column: tuple[int, int],
+    column: tuple[float, float],
 ) -> tuple[tuple[float, float], ...]:
     """Per-knob contact offsets under the StableLego *paper* rule.
